refactor: rename processor-api to plugin-api
