+++ injectors/cdi-injector/src/main/kotlin/io/mcarle/konvert/injector/cdi/CdiInjector.kt
@@ -7,8 +7,8 @@ import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.DelicateKotlinPoetApi
 import com.squareup.kotlinpoet.TypeSpec
-import io.mcarle.konvert.processor.api.KonverterInjector
-import java.lang.reflect.Proxy
+import io.mcarle.konvert.plugin.api.KonverterInjector
+import io.mcarle.konvert.plugin.api.extendProxy
 
 @AutoService(KonverterInjector::class)
 class CdiInjector : KonverterInjector {
@@ -29,21 +29,6 @@ class CdiInjector : KonverterInjector {
             builder.addAnnotation(
                 AnnotationSpec.get(it.value.extendProxy(), false)
             )
-        }
-    }
-
-    private inline fun <reified T : Any> T.extendProxy(): T {
-        return if (Proxy.isProxyClass(this::class.java)) {
-            val ih = Proxy.getInvocationHandler(this)
-            Proxy.newProxyInstance(T::class.java.classLoader, arrayOf(T::class.java)) { proxy, method, args ->
-                if (method.name == "annotationType") {
-                    T::class.java
-                } else {
-                    ih.invoke(proxy, method, args)
-                }
-            } as T
-        } else {
-            this
         }
     }
 }
